simd -> lmsd
